raspberrypi/requirements.txt
awsiotsdk
board
adafruit_dht